Rename project (#17)
* Rename project files

* Rename project in docs and CI workflows

* Set version to 1.0.0
