+++ .github/workflows/pull-request.yml
@@ -19,7 +19,7 @@ jobs:
         id: source-version
         run: |
           echo "Source branch: ${{ github.head_ref }}"
-          version=$(cat ./src/Dodo.HttpClientExtensions/Dodo.HttpClientExtensions.csproj | grep "<VersionPrefix>" | sed -e "s/ *<\/*VersionPrefix>//g")
+          version=$(cat ./src/Dodo.HttpClient.ResiliencePolicies/Dodo.HttpClient.ResiliencePolicies.csproj | grep "<VersionPrefix>" | sed -e "s/ *<\/*VersionPrefix>//g")
           echo "Source version: $version"
           if [ -z "${version// }" ]; then exit 1; fi
           echo "::set-output name=version::$version"
@@ -32,7 +32,7 @@ jobs:
         id: target-version
         run: |
           echo "Target branch: ${{ github.base_ref }}"
-          version=$(cat ./src/Dodo.HttpClientExtensions/Dodo.HttpClientExtensions.csproj | grep "<VersionPrefix>" | sed -e "s/ *<\/*VersionPrefix>//g")
+          version=$(cat ./src/Dodo.HttpClient.ResiliencePolicies/Dodo.HttpClient.ResiliencePolicies.csproj | grep "<VersionPrefix>" | sed -e "s/ *<\/*VersionPrefix>//g")
           echo "Source version: $version"
           if [ -z "${version// }" ]; then exit 1; fi
           echo "::set-output name=version::$version"
