+++ src/Dodo.HttpClient.ResiliencePolicies.Tests/Dodo.HttpClient.ResiliencePolicies.Tests.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Dodo.HttpClientExtensions\Dodo.HttpClientExtensions.csproj" />
+    <ProjectReference Include="..\Dodo.HttpClient.ResiliencePolicies\Dodo.HttpClient.ResiliencePolicies.csproj" />
   </ItemGroup>
 
 </Project>
