+++ src/Dodo.HttpClient.ResiliencePolicies.Tests/TimeoutPolicyTests.cs
@@ -1,10 +1,12 @@
 using System;
 using System.Net;
 using System.Threading.Tasks;
+using Dodo.HttpClient.ResiliencePolicies.RetrySettings;
+using Dodo.HttpClient.ResiliencePolicies.Tests.DSL;
 using NUnit.Framework;
 using Polly.Timeout;
 
-namespace Dodo.HttpClientExtensions.Tests
+namespace Dodo.HttpClient.ResiliencePolicies.Tests
 {
 	[TestFixture]
 	public class TimeoutPolicyTests
