+++ src/Dodo.HttpClient.ResiliencePolicies/CircuitBreakerSettings/ICircuitBreakerSettings.cs
@@ -2,7 +2,7 @@ using System;
 using System.Net.Http;
 using Polly;
 
-namespace Dodo.HttpClientExtensions
+namespace Dodo.HttpClient.ResiliencePolicies.CircuitBreakerSettings
 {
 	public interface ICircuitBreakerSettings
 	{
