+++ src/Dodo.HttpClient.ResiliencePolicies/RetrySettings/SimpleRetrySettings.cs
@@ -2,7 +2,7 @@
 using System.Net.Http;
 using Polly;
 
-namespace Dodo.HttpClientExtensions
+namespace Dodo.HttpClient.ResiliencePolicies.RetrySettings
 {
 	public class SimpleRetrySettings : IRetrySettings
 	{
